+++ examples/transtab/transtab.py
@@ -25,8 +25,8 @@ sys.path.append("../")
 sys.path.append("../../")
 from rllm.types import ColType
 from rllm.datasets import Titanic, Adult
-from rllm.nn.models import TransTabClassifier
 from rllm.preprocessing import TokenizerConfig
+from rllm.nn.models import TransTabClassifier
 import utils_run
 
 parser = argparse.ArgumentParser()
