+++ examples/transtab/utils_run.py
@@ -1,14 +1,14 @@
-from typing import Dict, List, Optional
-import time
 import os
 import random
+import time
+from typing import Dict, List, Optional
 
-from tqdm import tqdm
 import numpy as np
+from sklearn.metrics import roc_auc_score, accuracy_score, f1_score
 import torch
 import torch.nn.functional as F
 from torch.utils.data import DataLoader
-from sklearn.metrics import roc_auc_score, accuracy_score, f1_score
+from tqdm import tqdm
 
 
 __all__ = [
