+++ rllm/nn/models/__init__.py
@@ -16,6 +16,7 @@ from .tabnet import (
 )
 from .bridge import BRIDGE, TableEncoder, GraphEncoder
 from .transtab import TransTab, TransTabClassifier, TransTabForCL
+from .base_model import LinearClassifier
 
 __all__ = [
     "RECT_L",
@@ -26,4 +27,5 @@ __all__ = [
     "TransTab",
     "TransTabClassifier",
     "TransTabForCL",
+    "LinearClassifier",
 ]
